Move prover test methods
