+++ aries_vcx/tests/utils/devsetup_util.rs
@@ -1,13 +1,17 @@
-#[cfg(feature = "modular_libs")]
-use aries_vcx::core::profile::modular_libs_profile::ModularLibsProfile;
+use agency_client::agency_client::AgencyClient;
 use aries_vcx::core::profile::profile::Profile;
+use aries_vcx::errors::error::VcxResult;
+use aries_vcx::handlers::connection::mediated_connection::MediatedConnection;
+use aries_vcx::handlers::proof_presentation::mediated_prover;
+use aries_vcx::handlers::proof_presentation::prover::Prover;
+use aries_vcx::protocols::proof_presentation::prover::state_machine::ProverState;
 use aries_vcx_core::wallet::base_wallet::BaseWallet;
+use messages::msg_fields::protocols::present_proof::PresentProof;
+use messages::AriesMessage;
 
 #[cfg(test)]
 pub mod test_utils {
     use agency_client::api::downloaded_message::DownloadedMessage;
-    #[cfg(feature = "modular_libs")]
-    use aries_vcx::core::profile::modular_libs_profile::ModularLibsProfile;
     use aries_vcx::errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult};
     use messages::msg_fields::protocols::connection::Connection;
     use messages::msg_fields::protocols::cred_issuance::CredentialIssuance;
@@ -73,4 +77,38 @@ pub mod test_utils {
         }
         None
     }
+}
+
+pub async fn get_proof_request_messages(
+    agency_client: &AgencyClient,
+    connection: &MediatedConnection,
+) -> VcxResult<String> {
+    let presentation_requests: Vec<AriesMessage> = connection
+        .get_messages(agency_client)
+        .await?
+        .into_iter()
+        .filter_map(|(_, message)| match message {
+            AriesMessage::PresentProof(PresentProof::RequestPresentation(_)) => Some(message),
+            _ => None,
+        })
+        .collect();
+
+    Ok(json!(presentation_requests).to_string())
+}
+
+pub async fn prover_update_with_mediator(
+    sm: &mut Prover,
+    agency_client: &AgencyClient,
+    connection: &MediatedConnection,
+) -> VcxResult<ProverState> {
+    trace!("prover_update_with_mediator >>> ");
+    if !sm.progressable_by_message() {
+        return Ok(sm.get_state());
+    }
+    let messages = connection.get_messages(agency_client).await?;
+    if let Some((uid, msg)) = mediated_prover::prover_find_message_to_handle(sm, messages) {
+        sm.process_aries_msg(msg.into()).await?;
+        connection.update_message_status(&uid, agency_client).await?;
+    }
+    Ok(sm.get_state())
 }
